retouch role conversion
